Manual rebase Windows build compatibility

+++ src/playhead/src/sub_playhead.cpp
@@ -222,13 +222,13 @@ void SubPlayhead::init() {
             return rp;
         },
 
-        [=](duration_frames_atom atom) -> result<int> {
+        [=](duration_frames_atom atom) -> result<size_t> {
             if (up_to_date_) {
                 return full_timeline_frames_.size() ? full_timeline_frames_.size() - 1 : 0;
             }
             // not up to date, we need to get the timeline frames list from
             // the source
-            auto rp = make_response_promise<int>();
+            auto rp = make_response_promise<size_t>();
             request(caf::actor_cast<caf::actor>(this), infinite, source_atom_v)
                 .then(
                     [=](caf::actor) mutable {
