+++ src/utility/src/frame_rate_and_duration.cpp
@@ -25,9 +25,9 @@ double FrameRateDuration::seconds(const FrameRate &override) const {
 int FrameRateDuration::frames(const FrameRate &override) const {
     long int frames = 0;
     if (override.count()) {
-        frames = std::round(duration_ / override);
+        frames = (long)std::round(duration_ / override);
     } else if (rate_.count()) {
-        frames = std::round(duration_ / rate_);
+        frames = (long)std::round(duration_ / rate_);
     }
     return static_cast<int>(frames);
 }
